Refactor de l'organisation des fichiers et dossiers
