+++ front/src/components/error-alert.tsx
