+++ front/src/pages/dashboard.tsx
@@ -3,7 +3,7 @@ import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { useAuth } from "@/contexts/auth-context";
 import { cn } from "@/lib/utils";
-import { Logo } from "./logo";
+import { Logo } from "../components/logo";
 
 export function Dashboard({
   className,
@@ -16,7 +16,7 @@ export function Dashboard({
   };
 
   return (
-    <div className={cn("flex flex-col gap-6", className)} {...props}>
+    <div className={cn("flex flex-col gap-6 m-3", className)} {...props}>
       <div className="flex flex-col items-center gap-2">
         <Logo />
         <h1 className="text-xl font-bold">Connexion réussie</h1>
