+++ front/src/pages/login.tsx
@@ -5,10 +5,10 @@ import { cn } from "@/lib/utils";
 import { Button } from "@/components/ui/button";
 import { ApiError, loginUser } from "@/lib/api";
 import { useAuth } from "@/contexts/auth-context";
-import { FormHeader } from "./form/form-header";
-import { FormField } from "./form/form-field";
-import { PasswordGeneration } from "./form/password-generation";
-import { ErrorAlert } from "./common/error-alert";
+import { FormHeader } from "../components/form/header";
+import { FormField } from "../components/form/field";
+import { PasswordGeneration } from "../components/form/password-generation";
+import { ErrorAlert } from "../components/error-alert";
 
 interface LoginFormProps extends React.ComponentProps<"div"> {
   onRegisterClick?: () => void;
